// src/app/interfaces/additional.ts
export interface Additional {
    idAdditionalInformaction: any;
    countryOfBirth: any;
    countryOfResidence: any;
    nationality: any;
    civilState: any;
    maritalStatus: any;
    numberImss: any;
    scholarship: any;
    phone: any;
    mobile: any;
    otherPhone: any;
    nowResidenceSeniority: any;
    previousResidenceSeniority: any;
    address: any;
    external_number: any;
    interior_number: any;
    suburb: any;
    municipality: any;
    city: any;
    state: any;
    postal_code: any;
    country: any;
    creaditApplication: any;
}